@@ -14,7 +14,7 @@ void main() {
       await editor.pumpAndTap();
       await editor.tapButtonWithIcon(Icons.format_quote);
 
-      BlockNode block = editor.document.root.children.first;
+      BlockNode block = editor.document.root.children.first as BlockNode;
       expect(block.style.get(NotusAttribute.block), NotusAttribute.block.quote);
     });
   });
